refactor(styles): Refactor media query
