+++ tests/styles/abstractions/funcs.test.ts
